Rename RestoreFiles to GoldenRetriever
