+++ src/plugins/GoldenRetriever/IndexedDBStore.js
