+++ src/plugins/GoldenRetriever/ServiceWorker.js
