+++ src/plugins/GoldenRetriever/ServiceWorkerStore.js
