+++ src/plugins/GoldenRetriever/cleanup.js
